requirements.txt
numpy==1.18.5
Pillow==8.2.0
torch==1.10.1
torchvision==0.14.1

